+++ NEWS.md
@@ -212,6 +212,7 @@
 * Added the function linear_IRF to calculate linear impulse response functions based on a single regime. Bootstrapped confidence bounds are also available for models that impose linear autoregressive dynamics. 
 * Added a possibility to constrain the mixing weight parameters alphas to fixed constants.
 * Added a possibility to constrain the lambda parameters of structural models to fixed constant.
+* Added the function Pearson_residuals to calculate the standardized Pearson residuals or the raw residuals.
 * fitGSMVAR can now filter out (many of the) inappropriate estimates by setting filter_estimates=TRUE. 
 * fitGSMVAR can now estimate without parallel computing and printout by specifying use_parallel=FALSE.
 * Updated the vignette.